Make the cost argument of state constructors a float.

+++ src/search/grid/grid_state.cc
@@ -24,7 +24,7 @@ using namespace std;
  * \param x The x-coordinate of this state.
  * \param y The y-coordinate of this state.
  */
-GridState::GridState(GridWorld *d, const State *parent, int g, int x, int y)
+GridState::GridState(GridWorld *d, const State *parent, float g, int x, int y)
 	: State(d, parent, g), x(x), y(y)
 {
 	this->h = domain->get_heuristic()->compute(this);
